Use a more natural range of inputs for unit tests
The dummy inputs now ramp from -1 to 1 in the normalized coordinate of
the input values. This should mean the nets are tested in a more
natural range closer to the one they were trained in. The ramp
function should be identical to `np.linspace(-1, 1, n_inputs)`.

Also updated the target values in the unit tests.