Add utilities to test RNNs
Added the script lwtnn-test-keras-rnn.py to read in and test a keras
saved network, together with (optionally) a pair of text files
specifying the pattern to evaluate. The formatting is identical to
that expected by the feed forward tests.

Also added the ability for lwtnn-test-rnn to read in a pair of text
files. Again the formatting is identical to that expected by the feed
forward tests.