Make LSTM and GRU stateless
The implementation of the LSTM and GRU layers relied on internal state
vectors which were mutated during calls to `scan`. This prevented the
method from being declared `const`, but more importantly it wasn't
thread safe.

This commit makes calls `IRecurrentLayer::scan(...)` `const` which
forced a number of changes to the layers to make them thread safe.