Minor refactoring in json parser
Moved some code out of the top level json parsing function and into
smaller functions. Did this in preparation for graph parsing code
which will reuse a lot of these smaller functions.

Also added a second way of retrieving default values if the "defaults"
key is missing from the top level JSON file. In this case we'll look
in the "inputs" object. This should resolve #4.