Change Graph node containers to unordered sets
The Graph object was loosing track of the ordering of nodes when it
was constructed. This was bad, but replacing the vectors that
contained the nodes with unordered_set containers also got rid of a
few extra containers that were used in the constructor.

Code seems to work now for LightweightGraph. More validation will
follow.