Pass layer type to the layer converter file
This infomation is not available otherwise (all that's passed is the
parameters of the layer). So we add an args arguement to the other functions
and pass the layer_type as a 4th parameter.

The alternative is to have two seperate functions for GRU and LSTM and any
other rnn layers that may occur.