Remove an unused hack function
The `_get_merge_layer_parameters` function in the Keras layer
converters was originally introduced as a hack to support embedding in
sequential models. I've left it in the Keras 1.2 layer converters in
case we ever need to support that legacy case again, but it's been
removed from the Keras 2.0 converters.