Script to split Keras network into arch and weight
Our keras to json converters require that networks be saved as using a
json file and a weights component. Since most people save them using
the `model.save` method we often have to run a few lines to make the
conversion. This commit adds a script to make this easier.