Add scan option to lightweight graph
This adds the ability for the graph to return sequence outputs using
the `scan` method.

Since it's not possible to say what the graph will return when it is
constructed, we have to check at runtime. The network will throw a
OutputRankException if the specified output node returns a tensor of
the wrong rank.