Add basic unit test to check conversion process
This uses the lstm functional unit test data.
It then creates an input file and compares to the downloaded one.
It then creates the network file using the new input file and compares
to the old one.
In each case a diff is performed on new an old. If it's empty the test
is successful.