Allow empty input sequences
We originally threw exceptions when a sequence processing node
encountered an empty sequence. As mentioned in issue #80, graphs with
multiple inputs often have a good reason for an empty input sequence,
so we've relaxed this requirement.

To allow this we did have to make up a corner case assumption for
recurrent units with input sequences of length zero:

 - If we ask for the the _last_ vector in the sequence, we return a
   vector of zeros.

 - If we ask for the full sequence, we return a sequence of length zero.