Add support for alpha in ELu
Previously the ELu layer had a hardcoded alpha parameter. Since
several activation functions need more flexibility, the Activation
enum was promoted to a full structure in the C++ code.

For now the only activation that uses this is ELu. These changes are
still untested, so some unit test around ELu would be nice as well.