Look up LightweightGraph inputs lazily
In a graph with multiple outputs it's possible that only some inputs
will be needed to compute a given output. Prior to this commit,
LightweightGraph would eagerly convert all inputs into to the Eigen
representation before passing them to the Graph. This meant all of the
inputs had to exist in the LightweightGraph input map, even those
which the graph doesn't use.

This commit pushes the std::map to Eigen conversion into the Source
object, which means that the conversion only takes place if it's
required to compute the output in question.