Activation layers inherit rank from source
Previously the converter scripts would assume that any "activation"
layer in Keras was a feed-forward layer. This broke in cases where the
activation was operating on a sequence.

This commit makes all activation functions inherit their source node's
rank: i.e. if the source is a sequence, the activation function will
be time-distributed, if the source is feed-forward, the activation
will be feed-forward as well.