Fix for shared activation functions
One activation layer instance can be connected to a keras network in several
nodes. As an optimization, kerasfunc2json would recycle some node parameters
if it encountered the same layer twice. One of the recycled parameters was the
number of node outputs. This is usually fine: most layers have a fixed number
of outputs, but in the case of shared activation functions the number of
outputs doesn't need to be consistent. This was causing some conversions to
fail.

As a fix, I have removed the optimization: the number of outputs is calculated
for every node weather or not the layer is already constructed.