Templating important classes (#103)
Template the base type for all computation

Previously all the main classes responsible for computation used double precision. Now
the type is a template parameter, which allows us to experiment with any type supported
by Eigen. To keep backward compatibility the old classes are still in the same location and
behave the same way, but they are wrappers on the template versions (using `T = double`).
The template code lives in `include/lwtnn/generic`, and keeps the same class names but
prefixed with the the namespace `generic::`.

We introduced templates to support autodiff types, but we could also use them to reduce
the memory footprint of NNs by using `float` instead of `double`, or to experiment with 
other lower precision floating point formats.

Co-authored-by: Benjamin <benjamin@localhost.localdomain>
Co-authored-by: ductng <ductng@gmx.net>
Co-authored-by: Dan Guest <dguest@cern.ch>