Add a test for the cmake installation, fix cmake bugs
There were quite a few things missing if someone used CMake to install this
library. This should fix a few of them. I also asked the cmake build script to
install the library and then try to build another library and link against it,
so hopefully we can catch more of these issues in the future.